+++ Compatibility/ValheimPlusCompat.cs
@@ -1,5 +1,6 @@
 ﻿using HarmonyLib;
 using BepInEx.Bootstrap;
+using System.Reflection;
 
 namespace ExtraSlots.Compatibility;
 
@@ -7,41 +8,33 @@ internal static class ValheimPlusCompat
 {
     public const string valheimPlusGuid = "org.bepinex.plugins.valheim_plus";
 
-    public static bool VPlusInstalled = Chainloader.PluginInfos.ContainsKey(valheimPlusGuid);
-
-    internal static bool scrollBarSubstituted = false;
-
-    [HarmonyPatch(typeof(InventoryGui), nameof(InventoryGui.Show))]
-    public static class InventoryGui_Show_VPLusPrefix
+    internal static void CheckForCompatibility()
     {
-        [HarmonyBefore(valheimPlusGuid)]
-        [HarmonyPriority(Priority.First)]
-        public static void Postfix(InventoryGui __instance)
+        if (Chainloader.PluginInfos.ContainsKey(valheimPlusGuid))
         {
-            if (!VPlusInstalled)
-                return;
-
-            if (__instance.m_playerGrid.m_scrollbar == null)
+            // Unpatch redundant method that change inventory gui
+            MethodInfo method = AccessTools.Method("InventoryGui:Show");
+            MethodInfo patch = AccessTools.Method("ValheimPlus.GameClasses.InventoryGui_Show_Patch:Postfix");
+            if (method != null && patch != null)
             {
-                __instance.m_playerGrid.m_scrollbar = __instance.m_containerGrid.m_scrollbar;
-                scrollBarSubstituted = true;
+                ExtraSlots.instance.harmony.Unpatch(method, patch);
+                ExtraSlots.LogInfo("ValheimPlus.GameClasses.InventoryGui_Show_Patch:Postfix was unpatched.");
             }
-        }
-    }
 
-    [HarmonyPatch(typeof(InventoryGui), nameof(InventoryGui.Show))]
-    public static class InventoryGui_Show_VPLusPostfix
-    {
-        [HarmonyAfter(valheimPlusGuid)]
-        public static void Postfix(InventoryGui __instance)
-        {
-            if (!VPlusInstalled)
-                return;
-
-            if (scrollBarSubstituted)
+            method = AccessTools.Method("InventoryGrid:UpdateGui");
+            patch = AccessTools.Method("ValheimPlus.GameClasses.InventoryGrid_UpdateGui_Patch:Prefix");
+            if (method != null && patch != null)
             {
-                __instance.m_playerGrid.m_scrollbar = null;
-                scrollBarSubstituted = false;
+                ExtraSlots.instance.harmony.Unpatch(method, patch);
+                ExtraSlots.LogInfo("ValheimPlus.GameClasses.InventoryGrid_UpdateGui_Patch:Prefix was unpatched.");
+            }
+
+            ConstructorInfo ctor = AccessTools.Constructor(typeof(Inventory));
+            patch = AccessTools.Method("ValheimPlus.GameClasses.Inventory_Constructor_Patch:Prefix");
+            if (ctor != null && patch != null)
+            {
+                ExtraSlots.instance.harmony.Unpatch(ctor, patch);
+                ExtraSlots.LogInfo("ValheimPlus.GameClasses.Inventory_Constructor_Patch:Prefix was unpatched.");
             }
         }
     }